requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
line-bot-sdk>=3.14.0
anthropic>=0.40.0
pydub>=0.25.1
pydantic-settings>=2.7.0
python-multipart>=0.0.20
httpx>=0.28.0
pytest>=8.3.0
pytest-asyncio>=0.25.0
